@@ -14,20 +14,26 @@
     <name>eidas-idporten-connector</name>
     <description>Specific eidas-connector for Norway</description>
     <properties>
-        <java.version>24</java.version>
+        <java.version>25</java.version>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <open.telemetry.version>1.54.1</open.telemetry.version>
+        <eidas.version>2.9.0</eidas.version>
+        <open.telemetry.version>1.55.0</open.telemetry.version>
         <logstash.logback.version>8.1</logstash.logback.version>
         <idporten-audit.version>2.5.3</idporten-audit.version>
         <idporten-access-log.version>2.8.2</idporten-access-log.version>
+        <idporten-lib-keystore-core.version>1.1.1</idporten-lib-keystore-core.version>
         <logback-access.version>1.4.14</logback-access.version>
         <lombok.version>1.18.42</lombok.version>
+        <commons-lang.version>2.6</commons-lang.version>
+        <glassfish-jaxb.version>4.0.6</glassfish-jaxb.version>
+        <ibm-icu4j.version>77.1</ibm-icu4j.version>
         <idporten-metric-constants.version>1.1.0</idporten-metric-constants.version>
         <idporten-actuator-starter.version>2.0.0</idporten-actuator-starter.version>
         <nimbus.oauth2-oidc-sdk.version>11.30</nimbus.oauth2-oidc-sdk.version>
         <icu4j-countrycode-converter.version>78.1</icu4j-countrycode-converter.version>
         <surefire.plugin.version>3.5.4</surefire.plugin.version>
         <idporten-oidc-server-redis-cache-starter.version>1.0.6</idporten-oidc-server-redis-cache-starter.version>
+        <mockito-inline.version>5.2.0</mockito-inline.version>
         <maven.compiler-plugin.version>3.14.1</maven.compiler-plugin.version>
         <maven.compiler.proc>full</maven.compiler.proc>
     </properties>
