@@ -2,24 +2,25 @@ package no.idporten.eidas.connector.web;
 
 import com.nimbusds.oauth2.sdk.AuthorizationResponse;
 import com.nimbusds.oauth2.sdk.id.State;
-import com.nimbusds.openid.connect.sdk.Nonce;
 import eu.eidas.auth.commons.EIDASStatusCode;
 import eu.eidas.auth.commons.EidasParameterKeys;
+import eu.eidas.auth.commons.light.ILightRequest;
 import eu.eidas.auth.commons.light.ILightResponse;
 import jakarta.servlet.http.HttpServletRequest;
 import no.idporten.eidas.connector.config.StaticResourcesProperties;
 import no.idporten.eidas.connector.config.WebSecurityConfig;
-import no.idporten.eidas.connector.integration.specificcommunication.caches.CorrelatedRequestHolder;
 import no.idporten.eidas.connector.integration.specificcommunication.config.EidasCacheProperties;
-import no.idporten.eidas.connector.integration.specificcommunication.service.OIDCRequestStateParams;
 import no.idporten.eidas.connector.integration.specificcommunication.service.SpecificCommunicationService;
 import no.idporten.eidas.connector.logging.AuditService;
+import no.idporten.eidas.connector.service.AuthorizationResponseHelper;
 import no.idporten.eidas.connector.service.SpecificConnectorService;
 import no.idporten.eidas.lightprotocol.BinaryLightTokenHelper;
 import no.idporten.eidas.lightprotocol.IncomingLightResponseValidator;
-import no.idporten.eidas.lightprotocol.messages.*;
+import no.idporten.eidas.lightprotocol.messages.Attribute;
+import no.idporten.eidas.lightprotocol.messages.LevelOfAssurance;
+import no.idporten.eidas.lightprotocol.messages.LightResponse;
+import no.idporten.eidas.lightprotocol.messages.Status;
 import no.idporten.sdk.oidcserver.OpenIDConnectIntegration;
-import no.idporten.sdk.oidcserver.protocol.Authorization;
 import no.idporten.sdk.oidcserver.protocol.PushedAuthorizationRequest;
 import no.idporten.sdk.oidcserver.protocol.RedirectedResponse;
 import org.junit.jupiter.api.AfterEach;
@@ -66,6 +67,9 @@ class ConnectorResponseControllerTest {
     @MockitoBean
     private EidasCacheProperties eidasCacheProperties;
 
+    @MockitoBean
+    private AuthorizationResponseHelper authorizationResponseHelper;
+
     private final static String lightTokenId = "mockedLightTokenId";
 
     private static MockedStatic<BinaryLightTokenHelper> binaryLightTokenHelperMock;
@@ -93,29 +97,35 @@ class ConnectorResponseControllerTest {
         when(staticResourcesProperties.getStaticResourcesBaseUri()).thenReturn("http://static.idporten.no/latest/");
     }
 
+
     @Test
-    @DisplayName("then if there is a valid response return redirect to authorization endpoint")
-    void testValidLightRequest() throws Exception {
+    @DisplayName("then if there is a valid response returning UserMatchRedirect, controller redirects to given URL")
+    void testValidLightRequestUserMatchRedirect() throws Exception {
 
         LightResponse lightResponse = getLightResponse("relayState");
-        when(specificConnectorService.getCachedRequest(any(String.class))).thenReturn(new CorrelatedRequestHolder(LightRequest.builder().id("abc").relayState("relayState").build(),
-                new OIDCRequestStateParams(new State("123"),
-                        new Nonce("123"),
-                        null,
-                        "mockedTraceId"
-                )));
+        // Cached request with matching inResponseTo id
+        ILightRequest iLightRequest = mock(ILightRequest.class);
+        when(iLightRequest.getId()).thenReturn("abc");
+        no.idporten.eidas.connector.integration.specificcommunication.caches.CorrelatedRequestHolder holder =
+                new no.idporten.eidas.connector.integration.specificcommunication.caches.CorrelatedRequestHolder(
+                        iLightRequest,
+                        new no.idporten.eidas.connector.integration.specificcommunication.service.OIDCRequestStateParams(
+                                new State("123"), null, null, "traceId")
+                );
+
+        when(specificConnectorService.getCachedRequest(any(String.class))).thenReturn(holder);
         when(specificCommunicationService.getAndRemoveResponse(any(String.class), any())).thenReturn(lightResponse);
-        RedirectedResponse clientResponse = mock(RedirectedResponse.class);
-        when(clientResponse.toQueryRedirectUri()).thenReturn(new URI("http//junit?client_id=123&request_uri=http://redirect-url.com"));
-        PushedAuthorizationRequest authorizationRequest = mock(PushedAuthorizationRequest.class);
-        when(specificConnectorService.getAuthorization(lightResponse)).thenReturn(mock(Authorization.class));
-        when(openIDConnectSdk.authorize(eq(authorizationRequest), any(Authorization.class))).thenReturn(mock(no.idporten.sdk.oidcserver.protocol.AuthorizationResponse.class));
-        when(openIDConnectSdk.createClientResponse(any(no.idporten.sdk.oidcserver.protocol.AuthorizationResponse.class))).thenReturn(clientResponse);
+        when(specificConnectorService.matchUser(lightResponse))
+                .thenReturn(new no.idporten.eidas.connector.matching.domain.UserMatchRedirect("http://redirect.example/next"));
+
         mockMvc.perform(post("/ConnectorResponse")
-                        .sessionAttr(SESSION_ATTRIBUTE_AUTHORIZATION_REQUEST, authorizationRequest)
+                        .sessionAttr(SESSION_ATTRIBUTE_AUTHORIZATION_REQUEST, mock(PushedAuthorizationRequest.class))
                 )
-                .andExpect(redirectedUrl("http//junit?client_id=123&request_uri=http://redirect-url.com"));
+                .andExpect(redirectedUrl("http://redirect.example/next"));
+        verify(auditService).auditLightResponse(any(LightResponse.class), any());
     }
+
+
 
 
     @Test
@@ -134,6 +144,7 @@ class ConnectorResponseControllerTest {
                         .sessionAttr(SESSION_ATTRIBUTE_AUTHORIZATION_REQUEST, authorizationRequest)
                 )
                 .andExpect(redirectedUrl("http//junit?token=hei"));
+        verify(auditService).auditLightResponse(any(LightResponse.class), any());
     }
 
     private static LightResponse getLightResponse(String relayState) {
@@ -154,7 +165,6 @@ class ConnectorResponseControllerTest {
 
     @AfterEach
     void tearDown() {
-        verify(auditService).auditLightResponse(any(LightResponse.class), any());
         // Ensure that the static mocks are closed after each test
         binaryLightTokenHelperMock.close();
         incomingLightResponseValidatorMock.close();
